+++ .env
@@ -1,2 +1,2 @@
-# Create in https://www.allkeysgenerator.com/Random/Security-Encryption-Key-Generator.aspx
+# One alternative is Create in https://www.allkeysgenerator.com/Random/Security-Encryption-Key-Generator.aspx
 SECRET_KEY="my_secrets_key"
