+++ app/middlewares/jwt_bearer.py
@@ -1,6 +1,6 @@
 from fastapi.security import HTTPBearer
 from fastapi import Request, HTTPException
-from jwt_manager import create_token, validate_token
+from utils.jwt_manager import validate_token
 
 class JWTBearer(HTTPBearer):
     async def __call__(self, request: Request):
