+++ app/schemas/movie.py
@@ -1,10 +1,6 @@
 from pydantic import BaseModel, Field, validator
-from typing import Optional, Union
 import datetime
-from fastapi.security import HTTPBearer
-from fastapi import Request
-from routers.jwt_manager import validate_token
-from fastapi import HTTPException
+from typing import Union, List
 
 class MovieAPI(BaseModel):
     # id: Union[int, None] = Field(None, ge=1, le=2000)
@@ -54,24 +50,3 @@ class MovieAPI(BaseModel):
             }
         }
         # exclude = ['id']
-        
-        
-class User(BaseModel):
-    email: str
-    password: int
-    
-    class Config:
-        schema_extra = {
-            "example": {
-                "email": "admin@gmail.com",
-                "password": 12345,
-            }
-        }
-        
-        
-class JWTBearer(HTTPBearer):
-    async def __call__(self, request: Request):
-        auth = await super().__call__(request)
-        data = validate_token(auth.credentials)
-        if data['email'] != "admin@gmail.com":
-            raise HTTPException(status_code=403, detail="Credenciales son invalidas")
