+++ app/utils/data.py
@@ -1,3 +1,6 @@
+# Data usada antes de creacion de servidor SQLAlchemy
+# Se deja por motivos didacticos pero puede ser eliminada ya que no es usada
+
 films = [
     {
         'id': 1,
