--- requirement.txt
@@ -1,5 +1,0 @@
-fastapi==0.95.0
-uvicorn==0.21.1
-python-dotenv==1.0.0
-pyjwt==2.6.0
-SQLAlchemy==2.0.9
